requirements.txt
asgiref==3.4.1
certifi==2021.5.30
Django == 3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
pip==21.2.4
PyJWT==2.1.0
pytz==2021.1
setuptools==58.0.4
sqlparse==0.4.2
wheel==0.37.0